requirements.txt
Flask==3.0.0
pycryptodome==3.19.0
Werkzeug==3.0.1
psutil
matplotlib
